% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cextgr.R
\docType{data}
\name{cextgr}
\alias{cextgr}
\title{Contour segments seedling extension growth raw data.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 6701 rows and 102 columns.}
\usage{
cextgr
}
\description{
Raw extension growth data extracted from seesapmas11 master file.
}
\keyword{datasets}

