% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extgr.R
\docType{data}
\name{extgr}
\alias{extgr}
\title{Contour segments seedling extension growth data.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 32998 rows and 12 columns.
\itemize{

    \item ID: Unique plant identifier, links to other seedling data in segplants.
    \item EX: Extension growth measurements (cm).  The number of measurements recorded
during a given census varied, refer to detailed notes.
    \item YEAR: Year of measurement.
    \item YRREX: Indicator variable: 1 means data after this year can be trusted.  This
was determined by taking into account browsed/broken tops of plants and new leaders in
subsequent years.
    \item PEX: Indicator variable: 1 if the measurement is "partial growth", 
meaning the measurement was taken partway through that year.  Recorded in 1989, 1998, and
1999.
    \item NL: Indicator variable: 1 if a new leader emerged that year (this was only 
measured in 1999).
    \item TERM: Notes on the condition of the terminal leader (recorded in 1999, and 2000).
    \item AGE: Estimated total age of plants (only measured in 1988).
    \item MINAGE:  Estimated minimum age of plants (recorded in 1988, 1989, 1998, and 1999).
    \item D1RM: Distance remaining from the last positively identified extension growth to
the root collar of ABBA/PIRU seedlings (recorded in 1998 and 1999, see detailed notes).
    \item D2RM: Distance remaining after accounting for the minimal age of ABBA/PIRU
seedlings (recorded in 1998 and 1999, see detailed notes).
    \item ENOTE: Indicator variable for various conditions of terminal leader:
    \itemize{
       \item 1: New leader
       \item 2: Browse off, no ext. growth
       \item 3: Not browsed, just no growth.
       \item 4: Dead top
       \item 5: No dominant leader
    }
}}
\usage{
extgr
}
\description{
This dataset contains extension growth data for seedlings, measured in 1988, 1989,
1998, and 1999.  There was also some additional data collected in 2000 about the 
health status of some of the tagged seedlings.  Seedling information from this dataset
is linked to `segplants` by the ID variable.
}
\keyword{datasets}

