% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedsub.R
\name{seedsub}
\alias{seedsub}
\title{This package just contains data on seedlings and their associated 
substrates from Mt. Moosilauke.}
\description{
This package just contains data on seedlings and their associated 
substrates from Mt. Moosilauke.
}

