% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segplots.R
\docType{data}
\name{segplots}
\alias{segplots}
\title{Contour segments plot-level data.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 157 rows and 13 columns.
\itemize{
   \item PID: Plot identifier, this is determined by unique combinations of \code{CONTNAM} and \code{STPACE}.
   \item CONTNAM: Contour name
   \item STPACE: Beginning pace of a segment.
   \item ELEVCL: Elevation class
   \item ASPCL: Aspect class
   \item SOILCL[1-2]: Soil class
   \item HISTO: Soil type.
   \item SPODO: Soil type.
   \item INCEP: Soil type.
   \item MINER: ?
   \item ROCK: Soil type
   \item PHISTO: Soil type.
}}
\usage{
segplots
}
\description{
For more details, see \code{segdata}.  This data links to 
individual-level data in \code{segplants} by the `PID` variable.
}
\keyword{datasets}

