% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsub.R
\docType{data}
\name{segsub}
\alias{segsub}
\title{Contour segments substrate data}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 352 rows and 36 columns.
\itemize{
    \item PID: Plot identifier, links to `segplots`.  Some are NA, where only substrates
were measured.
    \item CONTNAM: Contour name
    \item STPACE: Pace along contour
    \item YEAR: Year of census.
    \item SGDSP: Segment displacement from contour ("U"=up, "D"=down) in paces (or m?)
    \item QPOS: Quadrat
    \item Substrate types:
    \itemize{
       \item BLA5/BLD5: Bole aerial/Bole dead
       \item BSOIL: Bare soil
       \item RCK: Rock
       \item WATER: Water
       \item WDG5/WDG1: Wood on the ground (at least 5cm/1cm diameter)
       \item MSSG: Moss
       \item TIPA: Tip-up
       \item STPA: ?
       \item LITT/LITM/LITC/LITD: General/Mixed/Coniferous/Deciduous litters
       \item WDA5/WDA1: Wood aerial (at least 5cm/1cm diameter)
       \item MSWDA5/MSWDA1: Moss on aerial wood (at least 5cm/1cm diameter)
       \item LITCRCK/LITMRCK/MSRCK: Coniferous/Mixed/Moss on rock
       \item MSBLA5: Moss on aerial bole (at least 5cm diameter)
       \item MSWDG5: Moss on dead aerial wood (at least 5cm diameter)
    }
    \item ASPCL: Aspect class
    \item ELEVCL: Elevation class
    \item SEASON: Season data was collected (only for summer of 2003)
    \item DATE: Date data was collected
    \item SUMG: Derived variable (sum on ground?)
    \item SUMA: Derived variable (sum of aerial?)
    \item CORRECT: Derived variable (?)
}}
\usage{
segsub
}
\description{
This dataset contains the substrate data for segment plots.  It was recorded
for different plots in 1989, 1998, 1999, and 2003.  Different protocols were followed
in different years, but the general idea was to record substrate as being 
moss, bare rock/soil, or various types of litter (mixed, deciduous, or coniferous).
}
\keyword{datasets}

